requirements.txt
SQLAlchemy~=2.0.15
python-dotenv~=1.0.0